pubmed_local_agent/requirements.txt
# # requirements.txt


# Core dependencies
openai>=1.3.0
python-dotenv==1.0.1
numpy==1.26.3
sentence-transformers==2.3.1
tqdm==4.66.2

# API and server
fastapi==0.109.2
uvicorn==0.27.1
python-multipart==0.0.9
gunicorn==21.2.0

# Vector store and document processing
faiss-cpu==1.7.4
pymupdf4llm~=0.0.21

# LangChain components
langchain-community>=0.0.20
langchain-core>=0.1.21

# Google Cloud dependencies
google-cloud-storage>=2.13.0

# Other utilities
requests>=2.31.0
scikit-learn>=1.2.0
anthropic>=0.5.0
typing-extensions>=4.7.0
xmltodict>=0.13.0
beautifulsoup4>=4.12.0
pandas>=2.0.0
tenacity>=8.2.0
matplotlib>=3.7.0
pydantic==2.6.1
backoff
tiktoken
